Refactor Editor
Squashed commit of the following:

commit e84094814ad1288bfa4fb192b776de56f3e75272
Author: blackening <blackening@gmail.com>
Date:   Tue Jun 9 23:53:03 2020 +0800

    Implement Typechecking for hotkey bindings.

commit d8a8bb23dcc4d9edc5659824277f78060121e06d
Author: blackening <blackening@gmail.com>
Date:   Tue Jun 9 13:43:42 2020 +0800

    Updated snapshot.

commit 9f37d071a2a05a7be983153f616005bc67c95634
Author: blackening <blackening@gmail.com>
Date:   Tue Jun 9 13:39:41 2020 +0800

    Formatting

commit 4407379bf6227cc3d7ba5446712fbfa925e255aa
Author: blackening <blackening@gmail.com>
Date:   Tue Jun 9 13:38:53 2020 +0800

    Remove all inline-comment stuff for PR.

commit df56093d27915998ec5b40fea8100abcb0f5d3b1
Author: blackening <blackening@gmail.com>
Date:   Tue Jun 9 12:53:48 2020 +0800

    Refactor TypeInference into mixin.

commit 94f8925086658c394cc140b58d1e30aca99e584f
Author: blackening <blackening@gmail.com>
Date:   Tue Jun 9 12:43:35 2020 +0800

    Impl mixins(Highlighting, Navigation, ShareAce)

commit 920bd49021ada2acbf9cf3e1bf3cf39f70319477
Author: blackening <blackening@gmail.com>
Date:   Tue Jun 9 10:05:06 2020 +0800

    Implementation of context menu.

commit 470877ae87d5551521dcde43731f894e6676537b
Author: blackening <blackening@gmail.com>
Date:   Mon Jun 8 11:32:39 2020 +0800

    Refactor Editor (Defaults, callbacks, typing)

commit 4fe3bd83a78147315b34282a534dbe15ed6f76fa
Author: blackening <blackening@gmail.com>
Date:   Mon Jun 8 01:16:03 2020 +0800

    Formatting.

commit d6a5ea553ccebb357385cf32c40e515ce493d1ef
Author: blackening <blackening@gmail.com>
Date:   Mon Jun 8 01:14:03 2020 +0800

    "Unstage" some changes

commit 94ddd5e118d8c949c991e46dc3b8ef5263d71142
Author: blackening <blackening@gmail.com>
Date:   Mon Jun 8 01:08:41 2020 +0800

    Formatting fixes

commit d01cf68c1272c864f2e3e5c2baa251a38dd529dd
Author: blackening <blackening@gmail.com>
Date:   Mon Jun 8 00:57:58 2020 +0800

    Refactor hotkey bindings + types.

commit 40945fad510e14873c9a99a36bfa09b8bdbfc073
Author: blackening <blackening@gmail.com>
Date:   Mon Jun 8 00:48:44 2020 +0800

    Abstract key-bindings into configuration file.

commit 77b36f73085f572533bba09fd354bca0df07a31f
Author: blackening <blackening@gmail.com>
Date:   Mon Jun 8 00:21:09 2020 +0800

    Updated typing for Editor.
