+++ src/commons/editor/Editor.tsx
@@ -1,27 +1,27 @@
-import { require as acequire } from 'ace-builds';
+/* tslint-disable no-unused-vars */
+import { require as acequire, Ace } from 'ace-builds';
 import 'ace-builds/src-noconflict/ext-language_tools';
 import 'ace-builds/src-noconflict/ext-searchbox';
 import * as React from 'react';
 import AceEditor, { IAceEditorProps } from 'react-ace';
 import { HotKeys } from 'react-hotkeys';
-import sharedbAce from 'sharedb-ace';
 
-import {
-  createContext,
-  getAllOccurrencesInScope,
-  getScope,
-  getTypeInformation,
-  hasDeclaration
-} from 'js-slang';
+import { createContext, getAllOccurrencesInScope } from 'js-slang';
 import { HighlightRulesSelector, ModeSelector } from 'js-slang/dist/editors/ace/modes/source';
 import 'js-slang/dist/editors/ace/theme/source';
 import { Variant } from 'js-slang/dist/types';
 
-import { checkSessionIdExists } from '../collabEditing/CollabEditingHelper';
 import { Documentation } from '../documentation/Documentation';
-import { Links } from '../utils/Constants';
 import AceRange from './EditorAceRange';
-import { Position } from './EditorTypes';
+import { AceMouseEvent, Position } from './EditorTypes';
+import { defaultKeyBindings as keyBindings, KeyBinding } from './HotkeyBindings';
+
+// =============== Mixins ===============
+import WithShareAce from './WithShareAce';
+import WithHighlighting from './WithHighlighting';
+import WithNavigation from './WithNavigation';
+import WithTypeInference from './WithTypeInference';
+export type Constructor<T> = new (...args: any[]) => T;
 
 /**
  * @property editorValue - The string content of the react-ace editor
@@ -60,29 +60,16 @@ type StateProps = {
   sourceVariant?: Variant;
 };
 
-class Editor extends React.PureComponent<EditorProps, {}> {
-  public ShareAce: any;
+export class EditorBase extends React.PureComponent<EditorProps, {}> {
   public AceEditor: React.RefObject<AceEditor>;
-  private markerIds: number[];
-  private onChangeMethod: (newCode: string) => void;
+  // HOC: These props will be injected into AceEditor on render
+  // Use this to pass any values/callbacks into it.
+  protected injectedRenderProps: { [key: string]: any } = {};
   private completer: {};
 
   constructor(props: EditorProps) {
     super(props);
     this.AceEditor = React.createRef();
-    this.ShareAce = null;
-    this.markerIds = [];
-    this.onChangeMethod = (newCode: string) => {
-      if (this.props.handleUpdateHasUnsavedChanges) {
-        this.props.handleUpdateHasUnsavedChanges(true);
-      }
-      this.props.handleEditorValueChange(newCode);
-      this.handleVariableHighlighting();
-      const annotations = this.AceEditor.current!.editor.getSession().getAnnotations();
-      if (this.props.isEditorAutorun && annotations.length === 0) {
-        this.props.handleEditorEval();
-      }
-    };
 
     this.completer = {
       getCompletions: (editor: any, session: any, pos: any, prefix: any, callback: any) => {
@@ -97,15 +84,19 @@ class Editor extends React.PureComponent<EditorProps, {}> {
     };
   }
 
+  public get editor() {
+    return this.AceEditor.current!.editor;
+  }
+
   public getBreakpoints() {
-    return this.AceEditor.current!.editor.session.getBreakpoints().filter(x => x != null);
+    return this.editor.session.getBreakpoints().filter(x => x != null);
   }
 
   public componentDidMount() {
     if (!this.AceEditor.current) {
       return;
     }
-    const editor = this.AceEditor.current!.editor;
+    const editor = this.editor;
     const session = editor.getSession();
 
     // TODO: Removal
@@ -135,28 +126,13 @@ class Editor extends React.PureComponent<EditorProps, {}> {
 
     // NOTE: the two `any`s below are because the Ace editor typedefs are
     // hopelessly incomplete
-    editor.on('gutterclick' as any, this.handleGutterClick);
+    editor.on('gutterclick' as any, this.handleGutterClick as any);
 
     // Change all info annotations to error annotations
     session.on('changeAnnotation' as any, this.handleAnnotationChange(session));
 
     // Start autocompletion
     acequire('ace/ext/language_tools').setCompleters([this.completer]);
-
-    // Has session ID
-    if (this.props.editorSessionId !== '') {
-      this.handleStartCollabEditing(editor);
-    }
-
-    this.handleVariableHighlighting();
-  }
-
-  public componentWillUnmount() {
-    if (this.ShareAce !== null) {
-      // Umounting, closing websocket
-      this.ShareAce.WS.close();
-    }
-    this.ShareAce = null;
   }
 
   public componentDidUpdate(prevProps: EditorProps) {
@@ -183,18 +159,9 @@ class Editor extends React.PureComponent<EditorProps, {}> {
 
   // chapter selector used to choose the correct source mode
   public chapterNo = () => {
-    let chapter = this.props.sourceChapter;
-    let variant = this.props.sourceVariant;
-    let external = this.props.externalLibraryName;
-    if (chapter === undefined) {
-      chapter = 1;
-    }
-    if (variant === undefined) {
-      variant = 'default';
-    }
-    if (external === undefined) {
-      external = 'NONE';
-    }
+    let chapter = this.props.sourceChapter || 1;
+    let variant = this.props.sourceVariant || 'default';
+    let external = this.props.externalLibraryName || 'NONE';
 
     HighlightRulesSelector(chapter, variant, external, Documentation.externalLibraries[external]);
     ModeSelector(chapter, variant, external);
@@ -207,48 +174,7 @@ class Editor extends React.PureComponent<EditorProps, {}> {
         <div className="row editor-react-ace">
           <AceEditor
             className="react-ace"
-            commands={[
-              {
-                name: 'evaluate',
-                bindKey: {
-                  win: 'Shift-Enter',
-                  mac: 'Shift-Enter'
-                },
-                exec: this.props.handleEditorEval
-              },
-              {
-                name: 'navigate',
-                bindKey: {
-                  win: 'Ctrl-B',
-                  mac: 'Command-B'
-                },
-                exec: this.handleNavigate
-              },
-              {
-                name: 'refactor',
-                bindKey: {
-                  win: 'Ctrl-M',
-                  mac: 'Command-M'
-                },
-                exec: this.handleRefactor
-              },
-              {
-                name: 'highlightScope',
-                bindKey: {
-                  win: 'Ctrl-Shift-H',
-                  mac: 'Command-Shift-H'
-                },
-                exec: this.handleHighlightScope
-              },
-              {
-                name: 'TypeInferenceDisplay',
-                bindKey: {
-                  win: 'Ctrl-Shift-M',
-                  mac: 'Command-Shift-M'
-                },
-                exec: this.handleTypeInferenceDisplay
-              }
-            ]}
+            commands={this.generateKeyBindings(keyBindings)}
             editorProps={{
               $blockScrolling: Infinity
             }}
@@ -258,8 +184,7 @@ class Editor extends React.PureComponent<EditorProps, {}> {
             height="100%"
             highlightActiveLine={false}
             mode={this.chapterNo()} // select according to props.sourceChapter
-            onChange={this.onChangeMethod}
-            onCursorChange={this.handleVariableHighlighting}
+            onChange={this.onChange.bind(this)}
             theme="source"
             value={this.props.editorValue}
             width="100%"
@@ -268,10 +193,23 @@ class Editor extends React.PureComponent<EditorProps, {}> {
               enableLiveAutocompletion: true,
               fontFamily: "'Inconsolata', 'Consolas', monospace"
             }}
+            {...this.injectedRenderProps}
           />
         </div>
       </HotKeys>
     );
+  }
+
+  protected onChange(newCode: string, delta: Ace.Delta) {
+    console.log('Change', newCode, delta);
+    if (this.props.handleUpdateHasUnsavedChanges) {
+      this.props.handleUpdateHasUnsavedChanges(true);
+    }
+    this.props.handleEditorValueChange(newCode);
+    const annotations = this.AceEditor.current!.editor.getSession().getAnnotations();
+    if (this.props.isEditorAutorun && annotations.length === 0) {
+      this.props.handleEditorEval();
+    }
   }
 
   // Used in navigating from occurence to navigation
@@ -282,55 +220,24 @@ class Editor extends React.PureComponent<EditorProps, {}> {
     this.AceEditor.current!.editor.renderer.scrollCursorIntoView(position, 0.5);
   };
 
-  private handleNavigate = () => {
-    const chapter = this.props.sourceChapter;
-    const variantString =
-      this.props.sourceVariant === 'default' ? '' : `_${this.props.sourceVariant}`;
-    const pos = this.AceEditor.current!.editor.selection.getCursor();
-    const token = this.AceEditor.current!.editor.session.getTokenAt(pos.row, pos.column);
-    const url = Links.textbook;
-
-    const external =
-      this.props.externalLibraryName === undefined ? 'NONE' : this.props.externalLibraryName;
-    const externalUrl =
-      this.props.externalLibraryName === 'ALL' ? `External%20libraries` : external;
-    const ext = Documentation.externalLibraries[external];
-
-    this.props.handleDeclarationNavigate(this.AceEditor.current!.editor.getCursorPosition());
-
-    const newPos = this.AceEditor.current!.editor.selection.getCursor();
-    if (newPos.row !== pos.row || newPos.column !== pos.column) {
-      return;
-    }
-
-    if (
-      hasDeclaration(this.props.editorValue, createContext(chapter), {
-        line: newPos.row + 1, // getCursorPosition returns 0-indexed row, function here takes in 1-indexed row
-        column: newPos.column
-      })
-    ) {
-      return;
-    }
-
-    if (ext.some((node: { caption: string }) => node.caption === (token && token.value))) {
-      if (
-        token !== null &&
-        (/\bsupport.function\b/.test(token.type) || /\bbuiltinConsts\b/.test(token.type))
-      ) {
-        window.open(`${url}source/${externalUrl}/global.html#${token.value}`); // opens external library link
+  private generateKeyBindings = (bindings: KeyBinding[]) => {
+    return bindings.map(cmd => {
+      const exec = typeof cmd.exec === 'function' ? cmd.exec : this[cmd.exec];
+      if (typeof exec !== 'function') {
+        console.error('Editor: Command cannot be bound due to invalid exec function', cmd, this);
+        throw new Error('Invalid Editor Command');
       }
-    } else if (
-      token !== null &&
-      (/\bsupport.function\b/.test(token.type) || /\bbuiltinconsts\b/.test(token.type))
-    ) {
-      window.open(`${url}source/source_${chapter}${variantString}/global.html#${token.value}`); // opens builtn library link
-    }
-    if (token !== null && /\bstorage.type\b/.test(token.type)) {
-      window.open(`${url}source/source_${chapter}.pdf`);
-    }
+      return { ...cmd, exec };
+    });
   };
 
-  private handleRefactor = () => {
+  // @ts-ignore. This is used by generateKeyBindings
+  public handleEditorEval = () => {
+    return this.props.handleEditorEval();
+  };
+
+  // @ts-ignore. This is used by generateKeyBindings
+  public handleRefactor = () => {
     const editor = (this.AceEditor.current as any).editor;
 
     if (!editor) {
@@ -352,118 +259,8 @@ class Editor extends React.PureComponent<EditorProps, {}> {
     ranges.forEach(range => selection.addRange(range));
   };
 
-  private handleHighlightScope = () => {
-    const editor = this.AceEditor.current!.editor;
-    if (!editor) {
-      return;
-    }
-    const code = this.props.editorValue;
-    const chapter = this.props.sourceChapter;
-    const position = editor.getCursorPosition();
-
-    const ranges = getScope(code, createContext(chapter), {
-      line: position.row + 1,
-      column: position.column
-    });
-
-    if (ranges.length !== 0) {
-      ranges.forEach(range => {
-        // Highlight the scope ranges
-        this.markerIds.push(
-          editor.session.addMarker(
-            new AceRange(
-              range.start.line - 1,
-              range.start.column,
-              range.end.line - 1,
-              range.end.column
-            ),
-            'ace_selection',
-            'text'
-          )
-        );
-      });
-    }
-  };
-
-  private handleVariableHighlighting = () => {
-    // using Ace Editor's way of highlighting as seen here: https://github.com/ajaxorg/ace/blob/master/lib/ace/editor.js#L497
-    // We use async blocks so we don't block the browser during editing
-
-    setTimeout(() => {
-      if (!this.AceEditor || !this.AceEditor.current || !this.AceEditor.current.editor) {
-        return;
-      }
-      const editor = (this.AceEditor.current as any).editor;
-      const session = editor.session;
-      const code = this.props.editorValue;
-      const chapterNumber = this.props.sourceChapter;
-      const position = editor.getCursorPosition();
-      if (!session || !session.bgTokenizer) {
-        return;
-      }
-      this.markerIds.forEach(id => {
-        session.removeMarker(id);
-      });
-      const ranges = getAllOccurrencesInScope(code, createContext(chapterNumber), {
-        line: position.row + 1,
-        column: position.column
-      }).map(
-        loc => new AceRange(loc.start.line - 1, loc.start.column, loc.end.line - 1, loc.end.column)
-      );
-
-      const markerType = 'ace_variable_highlighting';
-      const markerIds = ranges.map(range => {
-        // returns the marker ID for removal later
-        return session.addMarker(range, markerType, 'text');
-      });
-      this.markerIds = markerIds;
-    }, 10);
-  };
-
-  private handleTypeInferenceDisplay = (): void => {
-    const chapter = this.props.sourceChapter;
-    const code = this.props.editorValue;
-    const editor = this.AceEditor.current!.editor;
-    const pos = editor.getCursorPosition();
-    const token = editor.session.getTokenAt(pos.row, pos.column);
-
-    // comment out everyline of the inference string returned by getTypeInformation
-    const commentEveryLine = (str: string) => {
-      const arr = str.split('\n');
-      return arr
-        .filter(st => st !== '')
-        .map(st => '// ' + st)
-        .join('\n');
-    };
-
-    let output;
-
-    // display the information
-    if (this.props.handleSendReplInputToOutput) {
-      if (pos && token) {
-        // if the token is a comment, ignore it
-        if (token.type === 'comment') {
-          return;
-        }
-        const str = getTypeInformation(
-          code,
-          createContext(chapter),
-          { line: pos.row + 1, column: pos.column },
-          token.value
-        );
-        output = commentEveryLine(str);
-        if (str.length === 0) {
-          output = '// type information not found';
-        }
-      } else {
-        output = '// invalid token. Please put cursor on an identifier.';
-      }
-      this.props.handleSendReplInputToOutput(output);
-    }
-  };
-
-  private handleGutterClick = (e: any) => {
-    const target = e.domEvent.target;
+  private handleGutterClick = (e: AceMouseEvent) => {
+    const target = e.domEvent.target! as HTMLDivElement;
     if (
       target.className.indexOf('ace_gutter-cell') === -1 ||
       !e.editor.isFocused() ||
@@ -472,21 +269,23 @@ class Editor extends React.PureComponent<EditorProps, {}> {
       return;
     }
 
+    // Breakpoint related.
     const row = e.getDocumentPosition().row;
     const content = e.editor.session.getLine(row);
-    const breakpoints = e.editor.session.getBreakpoints(row, 0);
+    const breakpoints = e.editor.session.getBreakpoints();
     if (
       breakpoints[row] === undefined &&
       content.length !== 0 &&
       !content.includes('//') &&
       !content.includes('debugger;')
     ) {
+      // @ts-ignore: If breakpoint class not provided, it will be ace_breakpoint.
       e.editor.session.setBreakpoint(row);
     } else {
       e.editor.session.clearBreakpoint(row);
     }
     e.stop();
-    this.props.handleEditorUpdateBreakpoints(e.editor.session.$breakpoints);
+    this.props.handleEditorUpdateBreakpoints(e.editor.session.getBreakpoints());
   };
 
   private handleAnnotationChange = (session: any) => () => {
@@ -503,60 +302,6 @@ class Editor extends React.PureComponent<EditorProps, {}> {
       session.setAnnotations(annotations);
     }
   };
-
-  private handleStartCollabEditing = (editor: any) => {
-    const ShareAce = new sharedbAce(this.props.editorSessionId!, {
-      WsUrl: 'wss://' + Links.shareDBServer + 'ws/',
-      pluginWsUrl: null,
-      namespace: 'codepad'
-    });
-    this.ShareAce = ShareAce;
-    ShareAce.on('ready', () => {
-      ShareAce.add(
-        editor,
-        ['code'],
-        [
-          // TODO: Removal
-          // SharedbAceRWControl,
-          // SharedbAceMultipleCursors
-        ]
-      );
-      if (this.props.sharedbAceIsInviting) {
-        this.props.handleEditorValueChange(this.props.sharedbAceInitValue!);
-        this.props.handleFinishInvite!();
-      }
-    });
-
-    // WebSocket connection status detection logic
-    const WS = ShareAce.WS;
-    const sessionIdNotFound = () => {
-      clearInterval(interval);
-      WS.close();
-    };
-    const cannotReachServer = () => {
-      WS.reconnect();
-    };
-    const checkStatus = () => {
-      if (this.ShareAce === null) {
-        return;
-      }
-      checkSessionIdExists(
-        this.props.editorSessionId,
-        () => {},
-        sessionIdNotFound,
-        cannotReachServer
-      );
-    };
-    // Checks connection status every 5sec
-    const interval = setInterval(checkStatus, 5000);
-
-    WS.addEventListener('open', (event: Event) => {
-      this.props.handleSetWebsocketStatus!(1);
-    });
-    WS.addEventListener('close', (event: Event) => {
-      this.props.handleSetWebsocketStatus!(0);
-    });
-  };
 }
 
 /* Override handler, so does not trigger when focus is in editor */
@@ -564,15 +309,4 @@ const handlers = {
   goGreen: () => {}
 };
 
-// TODO: Removal
-// This interface is actually unused but ace poorly documents this feature so
-// we leave this here for reference.
-export interface IAutocompletionResult {
-  caption: string;
-  value: string;
-  meta?: string;
-  docHTML?: string;
-  score?: number;
-}
-
-export default Editor;
+export default WithTypeInference(WithNavigation(WithShareAce(WithHighlighting(EditorBase))));
